split the code to corresponding packages
