+++ packages/backend/src/rpc/index.ts
@@ -1,0 +1,3 @@
+export * from './serviceRegistry';
+export * from './dispatcher';
+export * from './events';
