+++ packages/backend/src/rpc/serviceRegistry.ts
@@ -1,0 +1,42 @@
+import * as bodyParser from 'body-parser';
+import * as Promise from 'bluebird';
+import { Request, Response } from 'express';
+import { App, BaseService, AppSingleton } from 'core';
+import { ContextualRouter, RPCDispatcher } from '..';
+
+export class RPCServiceRegistry extends AppSingleton {
+  private router = this.app.getSingleton(ContextualRouter);
+  private services: { [key: string]: typeof BaseService } = {};
+
+  constructor(app: App) {
+    super(app);
+    this.router.post('/rpc', bodyParser.json(), this.routeHandler.bind(this));
+  }
+
+  add(namespace: string, svc: typeof BaseService) {
+    if (this.services[namespace] != null) {
+      throw new Error('Namespace ' + namespace + ' already in use!');
+    }
+    this.services[namespace] = svc;
+  }
+
+  exists(serviceAlias: string, method: string) {
+    const ServiceClass = this.services[serviceAlias];
+    if (!ServiceClass) {
+      return false;
+    }
+    const serviceMethod = (ServiceClass.prototype as any)[method];
+    return typeof serviceMethod === 'function'; // && serviceMethod.__exposed;
+  }
+
+  get(serviceAlias: string) {
+    return this.services[serviceAlias];
+  }
+
+  routeHandler(req: Request, res: Response): void | Promise<void> {
+    let dispatcher = this.getSingleton(ContextualRouter)
+      .getContext(req, res)
+      .getService(RPCDispatcher);
+    return dispatcher.call();
+  }
+}
