+++ packages/example/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "uuid": "^3.2.1",
+    "core": "*",
     "backend": "*"
   },
   "devDependencies": {
