+++ packages/frontend/src/index.ts
@@ -1,1 +1,4 @@
-export const test = 'test';
+/**
+ * TODO gp: replace with frontend framework specifics.
+ */
+export const Router = 'placeholder';
